[feat] `cargo run deploy`
also had to add the following cargo subcommands
- `start`
- `stop`
